Separate parse_into_dtype util
