+++ py-polars/polars/io/csv/functions.py
@@ -13,8 +13,7 @@ from polars._utils.various import (
     normalize_filepath,
 )
 from polars._utils.wrap import wrap_df, wrap_ldf
-from polars.datatypes import N_INFER_DEFAULT, String
-from polars.datatypes.convert import py_type_to_dtype
+from polars.datatypes import N_INFER_DEFAULT, String, parse_into_dtype
 from polars.io._utils import (
     is_glob_pattern,
     parse_columns_arg,
@@ -501,7 +500,7 @@ def _read_csv_impl(
         if isinstance(schema_overrides, dict):
             dtype_list = []
             for k, v in schema_overrides.items():
-                dtype_list.append((k, py_type_to_dtype(v)))
+                dtype_list.append((k, parse_into_dtype(v)))
         elif isinstance(schema_overrides, Sequence):
             dtype_slice = schema_overrides
         else:
@@ -1218,7 +1217,7 @@ def _scan_csv_impl(
     if schema_overrides is not None:
         dtype_list = []
         for k, v in schema_overrides.items():
-            dtype_list.append((k, py_type_to_dtype(v)))
+            dtype_list.append((k, parse_into_dtype(v)))
     processed_null_values = _process_null_values(null_values)
 
     if isinstance(source, list):
